pyproject.toml
[tool.poetry]
name = "rpa_project"
version = "0.1.0"
description = ""
authors = ["Evgenii Smirnov <evgenii.smirnov@eficode.com>"]

[tool.poetry.dependencies]
python = "^3.10"
robotframework = "^5.0.1"
requests = "^2.28.1"
rpaframework = "^18.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
